extract tab activation handling and skip if browser is headless
